Changes
doesn't completely work, but it sets each color, then puts the hot
button as one of the 9 tiles. When it is clicked it says that it is the
hot button. Then it sets all tiles to white and then the hotbutton to
the desired color. until it runs out. Still haven't made it for the non
hotbutton, but you should get the point.